#97 - refactor: enhance sidebar components to utilize collection items and improve request handling - add folder management to collection store and update sidebar component
