+++ src/renderer/components/sidebar/SidebarRequestList/Nav/NavRequest.tsx
@@ -1,7 +1,4 @@
-import { RequestBodyType, TrufosRequest } from 'shim/objects/request';
-import { useCollectionStore, useCollectionActions } from '@/state/collectionStore';
-import { useEffect } from 'react';
-import { IpcPushStream } from '@/lib/ipc-stream';
+import { TrufosRequest } from 'shim/objects/request';
 import { SidebarMenuItem, SidebarMenuSubButton, SidebarMenuSubItem } from '@/components/ui/sidebar';
 import { RequestView } from '@/components/sidebar/SidebarRequestList/Nav/RequestView';
 
@@ -10,22 +7,6 @@ interface NavRequestProps {
 }
 
 export const NavRequest = ({ request }: NavRequestProps) => {
-  const { setSelectedRequest } = useCollectionActions();
-  const requestEditor = useCollectionStore((state) => state.requestEditor);
-  const selectedRequestIndex = useCollectionStore((state) => state.selectedRequestIndex);
-
-  useEffect(() => {
-    if (requestEditor == null) {
-      return;
-    } else if (request?.body?.type === RequestBodyType.TEXT && request?.id != null) {
-      IpcPushStream.open(request)
-        .then((stream) => IpcPushStream.collect(stream))
-        .then(requestEditor.setValue.bind(requestEditor));
-    } else {
-      requestEditor.setValue('');
-    }
-  }, [request?.id, requestEditor]);
-
   return (
     <SidebarMenuSubItem key={request.id}>
       <SidebarMenuSubButton asChild>
