--- src/renderer/components/sidebar/SidebarRequestList/index.css
@@ -1,9 +1,0 @@
-.sidebar-request-list-item {
-  color: #999999;
-  font-size: 14px;
-}
-
-.sidebar-request-list-item.selected {
-  background-color: #333333;
-  color: #EEEEEE;
-}
